Allow  config of whether to unit test min or max
By default unit tests run against minified code but devs generally
want to test unminified (max) code. Allow easy config switch.
